# docs/swagger.yaml
basePath: /api/v1/auth/
definitions:
  dto.SignInEmailPassword:
    properties:
      email:
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.SignUpEmailPassword:
    properties:
      contact_no:
        maxLength: 15
        minLength: 1
        type: string
      country_code:
        maxLength: 15
        minLength: 1
        type: string
      email:
        type: string
      first_name:
        maxLength: 50
        minLength: 1
        type: string
      last_name:
        maxLength: 50
        minLength: 1
        type: string
      locale:
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
    required:
    - contact_no
    - country_code
    - email
    - first_name
    - last_name
    - locale
    - password
    type: object
  models.ServiceStatus:
    properties:
      service:
        type: string
      status:
        type: string
      timestamp:
        type: integer
    type: object
  models.UserMeta:
    properties:
      contact_no:
        type: string
      country_code:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      locale:
        type: string
      updated_at:
        type: string
      verified_email:
        type: boolean
    type: object
host: localhost:9000
info:
  contact: {}
  description: OpenAPI Docs for Garm Auth Server
  title: Garm API Docs
  version: "1.0"
paths:
  /email-password/sign-in:
    post:
      consumes:
      - application/json
      description: Returns a JWT token for the user
      parameters:
      - description: Sign In Data
        in: body
        name: signUpDto
        required: true
        schema:
          $ref: '#/definitions/dto.SignInEmailPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserMeta'
      security:
      - ApiKeyAuth: []
      summary: Sign In with email and password
      tags:
      - email-password
  /email-password/sign-up:
    post:
      consumes:
      - application/json
      description: Creates a new user account using email and password
      parameters:
      - description: Sign Up Data
        in: body
        name: signUpDto
        required: true
        schema:
          $ref: '#/definitions/dto.SignUpEmailPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserMeta'
      security:
      - ApiKeyAuth: []
      summary: Sign up with email and password
      tags:
      - email-password
  /healthz:
    get:
      description: Returns the health status of the service including database connectivity
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ServiceStatus'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Check the health of the service
      tags:
      - health
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-api-token
    type: apiKey
swagger: "2.0"